Added a _cond_test inlined function to serve two purposes:
  Restrict conditional expressions to bool, classes, and integral
  types, as per the spec.

  Act on sync and single variables so they are read when used as
  conditional expressions.  This is not yet working because of an
  ambiguity resulting because a sync variable is itself a class and
  thus an object.

I use conditional expressions here to refer to those expressions
controlling if...then statements and expressions, while statements,
and do...while statements.

Added a test to show an error when using if on floats.  Previously
this "worked" and compared the float as C does.  Now it uses the
compilerError function to complain about the type.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@8943 3a8e244f-b0f2-452b-bcba-4c88e055c3ca