Updated lexer to work with cpkg/v5 (in particular with flex 2.5.33, bison 1.875d, and g++ 4.1.1).  Most of this involved renaming yylineno (a variable we had invented) because flex now defines it.  Also, dialed down warnings on generated C code for developers because the flex-generated code causes warnings.
Haven't yet checked to see whether this will fix the
problem that Wael was seeing (which he fixed in a
different manner).  That's work for next week.

This code compiled and ran under cpkg/v4 and cpkg/v5
on linux, and worked on the chap servers without
using modules, so hopefully won't break too much
this weekend.  :)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9007 3a8e244f-b0f2-452b-bcba-4c88e055c3ca