Categorized all .future files as follows:
? - The test needs to be examined and categorized.  If you are filing
    a future and are not positive what is wrong, label it "?" so that
    either you or someone else can look at it later as time allows.

semantic - The test illustrates something that is questionable in the
           language.  These need to be discussed as a group.  They
           should either be removed, categorized as a bug

error message - The test shows a place where the compiler should
                provide a better error message.

In addition to these general categories, the rest of the .futures now
have some specific bug category that summarizes what is wrong.  I only
use a user name to mean that this test is wrong and needs to be fixed
by this person.

As a proposal, I think we should use a username if you have taken
ownership of a bug or are working on it.  In this case, you can put
your user name next to the category in the future.  Here is the
breakdown of the 62 futures to date:

16 semantic
8 error message
4 generic iterators
3 global use before def
3 memory leak
3 uninstantiated generic parent
2 array of arrays
2 inout
1 ?
1 array fields and type aliases
1 array of syncs
1 cg-chpl-lineno
1 complicated expression in class
1 config var value in help message
1 const in iterator and param coercions
1 file I/O
1 formal type from default value
1 generic classes with default types
1 invalid executable command line arguments
1 lg2 error on real types
1 marybeth
1 nested classes
1 parameter folding
1 primitive to string conversions
1 promoted var function
1 query variables
1 return in type select
1 type in expression context
1 zippered iteration


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9060 3a8e244f-b0f2-452b-bcba-4c88e055c3ca