Expect regressions:
[Error matching program output for memory/shannon/freedWithoutMallocLineno]
[Error matching program output for memory/shannon/memmaxExceedsLineno]
[Error matching program output for memory/shannon/memstatFlagLineno]
[Error matching program output for memory/shannon/memtrackFlagLineno]
[Error matching program output for memory/shannon/reallocWithoutMallocLineno]
[Error matching program output for memory/shannon/outofmemory/mallocOutOfMemoryLineno]
[Error matching program output for trivial/shannon/readStringOverflowLineno]

Pair programmed the following with David.

Added primitives get_lineno and get_filename to mark places where user
line numbers and filenames should be printed.  This is currently only
done in the halt and assert functions.  It needs to be added to
certain primitives.

Removed global variables chpl_input_lineno and chpl_input_filename.

Fixed a .future where line numbers on errors were printed out
incorrectly in a parallel context.

Added a pass insertLineNumbers after resolution to insert line numbers
and filenames to functions that are in the standard modules so that
user line numbers and filename are printed out instead.

Removed access to global line number and filename from runtime.  This
leads to a few regressions that will be fixed when we start passing
line numbers and filenames to some primitives.  I will be working on
this now.

Updated output to compflags/bradc/help/userhelp.chpl.

Removed compiler option --cg-chpl-lineno and updated tests not to use
this.

Updated tests that didn't previously show line numbers because they
did not use --cg-chpl-lineno.  They now do show line numbers.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9119 3a8e244f-b0f2-452b-bcba-4c88e055c3ca