This checkin fixes the current regressions.
Added support for passing user code line number and filename
information to primitives in the runtime that require it for reporting
errors.  Added line number and filename arguments to many functions in
the runtime.  In some places in the runtime, I pass 0, 0 to avoid
adding it to more functions.  These need to be cleaned up in another
checkin.

Added PrimitiveOp::passLineno to signal that this primitive requires
line number and filename information.  This is added to the AST just
before code generation so as not to burden code in the rest of the
compiler.  Moved insertLineNumbers just before code generation.

Removed support for the "no codegen" pragma and turned the last of the
functions that used it (to test the runtime memory functions) into
primitives.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9128 3a8e244f-b0f2-452b-bcba-4c88e055c3ca