Another semantic future which breaks the "operations on arithmetic sequences" question from my previous check-in into its own test.  Fixing the CVSLOG message for that commit, it should have asked:
      These tests also raise the question of whether:

            (1..n)*2

      should result in the sequence:

            2, 3, 4, ... 2*n  (i.e., scale the low and high bounds)

      or:

            2, 4, 6, 8, 10, 2*n  (i.e., scale each sequence element)

This test more clearly illustrates that doing ops on
sequences results in the realization of the sequence,
without mixing it into the array slicing question
as in the previous test.  I'd like for us to consider
making arithmetic ops on scalars & arithmetic sequences
result in arithmetic sequences, so am filing this future
to help us confront that question (eventually).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9138 3a8e244f-b0f2-452b-bcba-4c88e055c3ca