Enabled writing to strings using a write method on strings as in
  s.write(A, b, i);

where s is a string and A, b, and i are variables of types that can be
written using the general write mechanism including the default write
methods that are created by the compiler.  This is not your typical
user-defineable string method as the 'this' argument is changed via
assignment.  Rationale: It is intuitive rather than orthogonal and the
oddness is contained.

Added support for inheriting from a Writer class that allows other
classes to be defined that can be written to.  The only method that
needs to be overridden is a method called "writeIt" that takes a
string.  (Yes, I am open to better names.)  To override the default
approach to writing for a given type, you now need to override a
method called writeThis (used to be write) that takes an argument of
type Writer.  An example is shown in

  classes/deitz/test_write_class.chpl

Updated the Input and Output section of the spec.

Added dtWriter to compiler to point to the Writer class type defined
in the standard modules.  Reworked the creation of the default
writeThis methods to use dtWriter in their definition.

Added pragma "ref this" to indicate that 'this' is a reference.  This
is used to implement the string.write method that requires changing
the implicit 'this' argument.

Fixed a bug in function resolution in which method and setter tokens
could be 'passed' to formal arguments of generic any types and create
functions instantiated by these tokens.

Changed all write methods in the modules and the tests that took files
to be writeThis methods that take Writers.

Added two test cases of writing to strings and one test case of a
user-defined Writer subclass.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9164 3a8e244f-b0f2-452b-bcba-4c88e055c3ca