Unified runtime and compiler runtime error message printing to use the same mechanism and style.
Removed get_lineno and get_filename primitives.  Added printError
primitive which is the function in the runtime that prints out runtime
errors.

Reimplemented and simplified insertLineNumbers to avoid the need for
the get_lineno and get_filename primitives.  This speeds it up by a
factor of two as well (though it is not a bottleneck).

Commented insertLineNumbers.

Reworked halt and assert in the standard modules to call into the
runtime error printing routine.  Used string.write functionality to
pass it a single string.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9167 3a8e244f-b0f2-452b-bcba-4c88e055c3ca