Added a distribution class hook.  There is no parent distribution class since the arrays are generic and this would violate return restrictions on class methods.  The Distribution class therefore must implement a structural interface.
Added a serial distribution class called

  SingleLocaleDistribution

that is the default distribution class.  It is used currently to build
our default arithmetic and associative domains.

Removed support for writing "distributed" and not specifying a
distribution.  Currently the distribution must be a value, not a type.

Defined a bogus variable Block to be an instance of the
SingleLocaleDistribution class to support its use in tests.

Changed reference to "block" distribution to refer to "Block" in
several of MB's tests.  Changed a reference to "blockcyclic" to
"Block" in studies/hpcc/RA/marybeth/ra-test-dist.chpl.

Expect major module code renaming and refactoring presently, and then
an example of defining a user-defined array and domain via a
distribution.  As always, I am open to naming suggestions, have only a
preference for consistency.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9175 3a8e244f-b0f2-452b-bcba-4c88e055c3ca