Fix a problem with the code: def foo(a:int, b:int, c:int) {   return (a, b, c); } var (a, _, c) = foo(1, 2, 3);
The _ in the tuple was being treated as a var name in the parser, so during
the checkParse phase an error was getting issued since "_" is not a valid
variable name.  Now throw away the value that would be stored to the "_"
position and don't complain about it or generate a variable for it.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9197 3a8e244f-b0f2-452b-bcba-4c88e055c3ca