Fix a problem with handling of "_" in tuple style variable defs.
Move the check for a symbol name of "_" to a point a little later in the
parse.  Previously a DefExpr was not inserted into the AST if its symbol name
was "_".  This threw off a count of which tuple field matched with which
variable, and resulted in a bad pointer dereference when the first field was
an underscore.  Now, the DefExpr is inserted, but later removed if its symbol
name is "_".


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9214 3a8e244f-b0f2-452b-bcba-4c88e055c3ca