Normalized returns and yields in preparation of moving the iterator transformation after function resolution.
Removed ReturnStmt subclass.  It is replaced by two primitives:
PRIMITIVE_RETURN and PRIMITIVE_YIELD.  This is more consistent within
our representation given the normalization of returns and yields such
that the return is the last statement in a function and has no effect
on code generation.

Though it should be clear that the elimination of unnecessary
variations is good, I will mention that this change both simplifies
the compiler and makes it more effective.  For example, without any
extra effort, copy propagation is more applicable and our generated
code is shorter.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9253 3a8e244f-b0f2-452b-bcba-4c88e055c3ca