Fix execution time segmentation violations when nested classes are used and the --unoptimized flag is on.
Was inserting an expression that did an assignment that looked like: "a = a;".
With optimization on, this got removed.  With optimization off, it got
translated to "free(a); a = a; touch(a);".

I stopped inserting the "a = a"; assignment, and flattened one other assignment
to use a temp variable, ie: "a = b;" became: "_tmp = b; a = _tmp;"


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9277 3a8e244f-b0f2-452b-bcba-4c88e055c3ca