Fixed the double allocation of arrays removing the three failures I mentioned in my last checkin.
Added new array type record called ArrayTypeInfo that is meant to
serve as runtime type information of an array.  The array type in
Chapel is complicated because the domain that is tied to the type has
runtime information associated with it that needs to be used to
declare an array of a given type.

Work is still required to enable arrays of arrays in a sane way, but
this may be a start.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9325 3a8e244f-b0f2-452b-bcba-4c88e055c3ca