Added a test case of a code that would produce a different answer if we inserted sequence temporaries for promotion.  In this code
  var n: int;

  def foo(i: int) {
    n = n + 1;
    writeln(i);
  }

  def bar(i: int) return n + i;

  foo(bar(1..10));

the call to bar promotes it over the values one to ten.  If the result
of the promotion of bar is captured in a sequence at this point, then
it returns the values one to ten exactly.  That is what currently
happens and the .good file is defined as such.  If a sequence does not
capture the results of bar but rather the results are sent into foo to
promote it, then n will be changed and the results that bar produces
will be changed as its execution is interleaved with foo.

I think this is how we should define the semantics of promotion and
forall expressions.  If we think of these as being possible in
parallel, then the semantics can be defined as potentially serial, in
an arbitrary order, or concurrently.  That would make the output
nondeterministic, but allows for parallelization and optimization.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9332 3a8e244f-b0f2-452b-bcba-4c88e055c3ca